Implemented the database in some methods (in CarBrandController). Fix MakeDate method in CarController. Fixed bugs.
