+++ CarDealership/Controllers/CarController.cs
@@ -18,21 +18,18 @@ namespace CarDealership.Controllers
         /// </summary>
         /// <param name="date">format=M.yyy :"10.2003"</param>
         /// <returns>DateTime with only moth and year</returns>
-        public static DateTime MakeDate(string date, string id)
+        public static DateTime MakeDate(string date)
         {
             try
             {
                 string[] dateArray = date.Split('.');//split the month and year
-                DateTime dateTime = new DateTime();//empty DateTime =1.1.0001
-                dateTime = dateTime.AddMonths(Convert.ToInt32(dateArray[0]) - 1);//add the months without the first
-                dateTime = dateTime.AddYears(Convert.ToInt32(dateArray[1]) - 1);//add the years without the first
                 
-                //add manufacture date to the car table 
-                using (carContext = new CarContext())
-                {
-                    var newDate = carContext.cars.Where(d => d.id == id).Select(d=> d.manufDate == dateTime);
-                    carContext.SaveChanges();
-                }
+                DateTime dateTime = new DateTime();//empty DateTime =1.1.0001
+                
+                dateTime = dateTime.AddMonths(Convert.ToInt32(dateArray[0]) - 1);//add the months without the first
+                
+                dateTime = dateTime.AddYears(Convert.ToInt32(dateArray[1]) - 1);//add the years without the first
+           
                 return dateTime;
             }
             catch (FormatException e)
@@ -64,7 +61,7 @@ namespace CarDealership.Controllers
         }
 
         /// <summary>
-        /// Show Cars in the Customer's Wishlist
+        /// Show Cars in the Customer's Wishlist    
         /// </summary>
         /// <param name="customerId">the user id using the app</param>
         /// <returns>list of ids</returns>
